Update based on review comments. Also a couple more in xx_decompose

+++ qiskit/transpiler/passes/synthesis/aqc_plugin.py
@@ -118,7 +118,7 @@ class AQCSynthesisPlugin(UnitarySynthesisPlugin):
         from qiskit.synthesis.unitary.aqc.cnot_unit_circuit import CNOTUnitCircuit
         from qiskit.synthesis.unitary.aqc.fast_gradient.fast_gradient import FastCNOTUnitObjective
 
-        num_qubits = int(round(math.log2(unitary.shape[0])))
+        num_qubits = round(math.log2(unitary.shape[0]))
 
         config = options.get("config") or {}
